dtraced: Create a cleanup action in dtraced.
This is activated by `dtracedctl --cleanup-state` currently. It's mostly a
workaround for a race condition which happens when dtraced forks a process
in the guest, but doesn't get the time to respond to the host dtrace(1)
process with its pid.

This results in the guest continuing to run a process that no longer has
anything listening for its DTrace state on the host, or a way to kill it.
As a workaround currently, we allow the user to simply run `dtracedctl` to
address various issues like this.

+++ cddl/contrib/opensolaris/cmd/dtraced/dtraced_state.c
@@ -184,6 +184,12 @@ init_state(struct dtd_state *s, int ctrlmachine, int nosha, int n_threads)
 	}
 
 	if ((err = mutex_init(
+	    &s->pidlistmtx, NULL, "pidlist", CHECKOWNER_YES)) != 0) {
+		dump_errmsg("Failed to create pidlist mutex: %m");
+		return (-1);
+	}
+
+	if ((err = mutex_init(
 	    &s->killcvmtx, NULL, "", CHECKOWNER_NO)) != 0) {
 		dump_errmsg("Failed to create kill condvar mutex: %m");
 		return (-1);
